Refactor Brave Ads mock utility and fix missing NotifyPrefDidChange
